instage/instage.service/requirements.txt
async-timeout==4.0.3
PyDMX==0.1.0
PyDMX-Drivers-FTDI==0.1.0
pylibftdi==0.21.0
redis==5.0.0
